requirements.txt
Pillow>=10.0.0
imagehash>=4.3.1
click>=8.1.0
tqdm>=4.64.0
numpy>=1.21.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0